# backend/src/main/java/com/stefvisser/springyield/controllers/AccountController.java
package com.stefvisser.springyield.controllers;

import com.stefvisser.springyield.dto.*;
import com.stefvisser.springyield.models.*;
import com.stefvisser.springyield.services.AccountService;
import com.stefvisser.springyield.services.UserService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.server.ResponseStatusException;

import java.util.List;
import java.util.Objects;

/**
 * REST controller for managing bank accounts.
 * <p>
 * This controller provides endpoints for retrieving, searching, and managing bank accounts.
 * It handles HTTP requests related to account operations and delegates the business logic
 * to the appropriate service classes. All endpoints are mapped to the "/api/account" base path.
 * </p>
 *
 * @since 1.0
 */
@RestController
@RequestMapping("/api/account")
public class AccountController {

    private final AccountService accountService;
    private final UserService userService;

    /**
     * Constructs a new AccountController with the required services.
     *
     * @param accountService service for account-related operations
     * @param userService service for user-related operations
     */
    public AccountController(AccountService accountService, UserService userService) {
        this.accountService = accountService;
        this.userService = userService;
    }

    /**
     * Searches for accounts based on specified criteria with pagination.
     * <p>
     * This endpoint allows filtering accounts by a search query and account type,
     * with results returned in a paginated format for efficient data transfer.
     * </p>
     *
     * @param query optional search string to filter accounts by various attributes
     * @param accountType optional account type filter (e.g., CURRENT, SAVINGS)
     * @param limit maximum number of results per page (defaults to 10)
     * @param offset starting position for pagination (defaults to 0)
     * @return ResponseEntity containing paginated account search results
     */
    @GetMapping("/search")
    public ResponseEntity<?> getAccountsBySearch(@AuthenticationPrincipal User user, @RequestParam(required = false) String query, @RequestParam(required = false) AccountType accountType, @RequestParam(required = false) AccountStatus status, @RequestParam(required = false) Integer limit, @RequestParam(required = false) int offset) {
        try {
            if (user == null)
                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User not authenticated");

            User executingUser = userService.getUserById(user.getUserId());

            if (!executingUser.isEmployee())
                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to view accounts");


            if (limit == null) limit = 10;
            if (offset < 0) offset = 0;
            if (query == null) query = "";

            PaginatedDataDTO<AccountProfileDto> accounts = accountService.searchAccount(query, accountType, status, limit, offset);
            return ResponseEntity.ok(accounts);
        } catch (ResponseStatusException e) {
            return ResponseEntity.status(e.getStatusCode()).body(e.getBody());
        }
    }

    @PutMapping("/{accountId}/limits")
    public ResponseEntity<?> updateAccountLimits(
            @AuthenticationPrincipal User user,
            @PathVariable Long accountId,
            @RequestBody AccountLimitsDTO limitsDTO) {
        try {
            if (user == null)
                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User not authenticated");

            User executingUser = userService.getUserById(user.getUserId());

            if (!executingUser.isEmployee()) {
                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to update account limits");
            }

            Account updatedAccount = accountService.updateAccountLimits(
                    accountId,
                    limitsDTO.getDailyLimit(),
                    limitsDTO.getAbsoluteLimit()
            );

            return ResponseEntity.ok(AccountProfileDto.wrap(updatedAccount));

        } catch (ResponseStatusException e) {
            return ResponseEntity.status(e.getStatusCode()).body(e.getBody());
        }
    }

    @GetMapping("/addressbook/search")
    public ResponseEntity<List<AddressBookEntryDTO>> searchAccountsByName(@RequestParam(required = false) String query) {
        PaginatedDataDTO<AccountProfileDto> accounts = accountService.searchAccountByName(query);

        if (accounts == null || accounts.getData().isEmpty()) {
            return ResponseEntity.notFound().build();
        }

        List<AddressBookEntryDTO> addressBookEntries = accounts.getData().stream()
                .map(AddressBookEntryDTO::wrap)
                .toList();

        return ResponseEntity.ok(addressBookEntries);
    }

    /**
     * Retrieves a user's profile along with their associated accounts.
     * <p>
     * This endpoint returns combined data about a user and their bank accounts,
     * providing a comprehensive view of a customer's profile and financial accounts.
     * </p>
     *
     * @param userId the unique identifier of the user
     * @return ResponseEntity containing a map with the user profile and their accounts
     */
    @GetMapping("/{userId}")
    public ResponseEntity<?> getUserAndAccounts(@AuthenticationPrincipal User user, @PathVariable Long userId) {
        try {
            if (user == null)
                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User not authenticated");

            User executingUser = userService.getUserById(user.getUserId());

            if (!executingUser.getUserId().equals(user.getUserId()) && !executingUser.isEmployee())
                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to view this user's profile");


            User foundUser = userService.getUserById(userId);

            if (foundUser == null) {
                return ResponseEntity.notFound().build();
            }

            return ResponseEntity.ok(new UserProfileDto(foundUser));
        } catch (ResponseStatusException e) {
            return ResponseEntity.status(e.getStatusCode()).body(e.getBody());
        }
    }

    @GetMapping("/iban/{iban}")
    public ResponseEntity<?> getAccountByIban(@AuthenticationPrincipal User user, @PathVariable String iban) {
        try {
            if (user == null)
                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User not authenticated");

            User executingUser = userService.getUserById(user.getUserId());

            Account account = accountService.getAccountByIban(iban);

            if (account == null) {
                return ResponseEntity.notFound().build();
            }

            if (!Objects.equals(account.getUser().getUserId(), executingUser.getUserId()) && !executingUser.isEmployee()){
                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to view this account");
            }

            return ResponseEntity.ok(new AccountProfileDto(account));
        } catch (ResponseStatusException e) {
            return ResponseEntity.status(e.getStatusCode()).body(e.getBody());
        }
    }
}