# backend/src/main/resources/static/api-docs.yaml
openapi: 3.0.0
info:
  title: Springyield API
  version: v0.1
servers:
  - url: http://springyield-springboot:8080
  - url: https://api.springyield.stefvisser.com
paths:
...
  /api/account/{accountId}/limits:
    put:
      tags:
        - account-controller
      operationId: updateAccountLimits
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountLimitsDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/auth/signup:
    post:
      tags:
        - auth-controller
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
  /api/auth/login:
    post:
      tags:
        - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
  /api/user/{userId}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/account/{userId}:
    get:
      tags:
        - account-controller
      operationId: getUserAndAccounts
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/account/search:
    get:
      tags:
        - account-controller
      operationId: getAccountsBySearch
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: accountType
          in: query
          required: false
          schema:
            type: string
            enum:
              - PAYMENT
              - SAVINGS
              - ATM
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - DEACTIVATED
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/account/iban/{iban}:
    get:
      tags:
        - account-controller
      operationId: getAccountByIban
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/account/addressbook/search:
    get:
      tags:
        - account-controller
      operationId: searchAccountsByName
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressBookEntryDTO'
...
components:
  schemas:
...
    AccountLimitsDTO:
      type: object
      properties:
        dailyLimit:
          type: number
        absoluteLimit:
          type: number
...
    Account:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        iban:
          type: string
        registrationDate:
          type: string
          format: date
        accountType:
          type: string
          enum:
            - PAYMENT
            - SAVINGS
            - ATM
        dailyLimit:
          type: number
        absoluteLimit:
          type: number
        balance:
          type: number
        balanceLimit:
          type: number
        status:
          type: string
          enum:
            - ACTIVE
            - DEACTIVATED
    AccountProfileDto:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        iban:
          type: string
        user:
          $ref: '#/components/schemas/User'
        accountType:
          type: string
          enum:
            - PAYMENT
            - SAVINGS
            - ATM
        balance:
          type: number
        dailyLimit:
          type: number
        absoluteLimit:
          type: number
        status:
          type: string
          enum:
            - ACTIVE
            - DEACTIVATED
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        bsnNumber:
          type: integer
          format: int32
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - UNAPPROVED
            - APPROVED
            - EMPLOYEE
            - ADMIN
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        authorities:
          type: array
          items:
            $ref: '#/components/schemas/GrantedAuthority'
        username:
          type: string
        employee:
          type: boolean
        approved:
          type: boolean
        enabled:
          type: boolean
        accountNonExpired:
          type: boolean
        credentialsNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
    UserSignupDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        bsnNumber:
          type: integer
          format: int32
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountProfileDto'
        approvedUser:
          type: boolean
    AuthResponseDto:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserProfileDto'
    UserProfileDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        bsnNumber:
          type: integer
          format: int32
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - UNAPPROVED
            - APPROVED
            - EMPLOYEE
            - ADMIN
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountProfileDto'
    UserLoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    AddressBookEntryDTO:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        iban:
          type: string